remove project.ci
The weird magic going on in the ci-{boost,build} scripts is honestly too
weird. With hindsight, it seems to me that it's much better to just
build a project with the same command during a CI run as when developing
locally.

Plus, I haven't really used either Travis or AppVeyor in quite some
time, so this code was mostly untested really.